fix(infra): move lambda to nested lambdas stack
refs. metriport/metriport-internal#2758
